Separate dialog cration function
